requirements.txt
api==0.0.7
appnope==0.1.0
argon2==0.1.10
argon2-cffi==19.1.0
aspy.yaml==1.2.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
blessings==1.7
boto3==1.9.93
botocore==1.12.142
cachetools==3.1.0
certifi==2019.3.9
cffi==1.12.3
cfgv==1.6.0
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
decorator==4.4.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.1.7
django-allauth==0.39.1
django-annoying==0.10.4
django-configurations==2.1
django-environ==0.4.5
django-filter==2.1.0
django-model-utils==3.1.2
django-nose==1.4.6
django-rest-swagger==2.2.0
django-storages==1.7.1
django-unique-upload==0.2.1
djangorestframework==3.9.1
docutils==0.14
entrypoints==0.3
environ==1.0
et-xmlfile==1.0.1
factory-boy==2.12.0
Faker==1.0.5
flake8==3.7.5
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
gunicorn==19.9.0
httplib2==0.12.3
identify==1.4.3
idna==2.8
importlib-metadata==0.11
ipdb==0.11
ipython==7.2.0
ipython-genutils==0.2.0
itypes==1.1.0
jdcal==1.4.1
jedi==0.13.3
Jinja2==2.10.1
jmespath==0.9.4
livereload==2.6.0
Markdown==3.0.1
MarkupSafe==1.1.1
mccabe==0.6.1
mkdocs==1.0.4
mock==2.0.0
more-itertools==7.0.0
newrelic==4.12.0.113
nodeenv==1.3.3
nose==1.3.7
nose-progressive==1.5.2
numpy==1.16.4
oauth2client==4.1.3
oauthlib==3.0.1
openapi-codec==1.3.2
openpyxl==2.6.2
pandas==0.24.2
parso==0.4.0
pbr==5.2.0
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.11.0
pre-commit==1.16.1
prompt-toolkit==2.0.9
psycopg2-binary==2.7.7
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
pytest==4.4.2
pytest-django==3.4.8
pytest-spec==1.1.0
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
PyYAML==5.1
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
s3transfer==0.2.0
simplejson==3.16.0
six==1.12.0
text-unidecode==1.2
toml==0.10.0
tornado==6.0.2
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.24.3
virtualenv==16.5.0
wcwidth==0.1.7
xlrd==1.2.0
zipp==0.5.0
